ominitags_scheduler/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --resolver=backtracking requirements.in
#
requests~=2.26.0
tortoise-orm==0.21.3
aiomysql==0.2.0